pyproject.toml
[tool.poetry]
name = "nonebot-plugin-nekoimage"
version = "0.3.0"
description = "A nonebot plugin for integrating with NekoImageGallery"
license = "AGPL-3.0"
authors = ["pk5ls20 <pk5ls20@outlook.com>"]
readme = "README.MD"
homepage = "https://github.com/pk5ls20/nonebot-plugin-nekoimage"
repository = "https://github.com/pk5ls20/nonebot-plugin-nekoimage"
documentation = "https://github.com/pk5ls20/nonebot-plugin-nekoimage/wiki"
keywords = ["nonebot", "nonebot2"]
packages = [{ include = "nonebot-plugin-nekoimage"}]

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.0.0"
pydantic = "*"
httpx = "*"
pytz = "*"
nonebot-plugin-send-anything-anywhere = "^0.4.0"
nonebot-adapter-onebot = "^2.0.0-beta.1"
nonebot-adapter-qq = "^1.3.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"